+++ CHANGELOG.md
@@ -7,3 +7,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 ### Added
 * Add documentation files: README.md; LICENSE; CHANGELOG.md; and, .gitignore.
+* Add bubble sort algorithm, implemented in Python.